# openapi/dist/aep-v1-m1.bundle.yaml
openapi: 3.1.0
info:
  title: Agent Enterprise Protocol M1
  version: 1.0.0-m1
  description: M1 contract for model discovery, authorization, and gateway connection.
  license:
    name: Apache 2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local M1 deployment
  - url: https://enterprise.example.com
    description: Protected deployment
security:
  - bearerAuth: []
...
paths:
...
  /aep/v1/user/skills/{skillId}/versions/{version}/package:
    get:
      tags:
        - User Runtime
      operationId: downloadSkillPackage
      parameters:
        - $ref: '#/components/parameters/SkillId'
        - $ref: '#/components/parameters/Version'
      responses:
        '200':
          description: Skill ZIP package
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/Problem-2'
        '404':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/user/skills/sync-results:
    post:
      tags:
        - User Runtime
      operationId: reportSkillSyncResult
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillSyncResult'
      responses:
        '202':
          description: Result accepted
        '400':
          $ref: '#/components/responses/Problem-2'
...
  /aep/v1/admin/skills:
    get:
      tags:
        - Admin Skills
      operationId: listSkills
      responses:
        '200':
          description: Skill list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSkillList'
    post:
      tags:
        - Admin Skills
      operationId: createSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillWrite'
      responses:
        '201':
          $ref: '#/components/responses/AdminSkill'
        '409':
          $ref: '#/components/responses/Problem-2'
  /aep/v1/admin/skills/{skillId}:
    parameters:
      - $ref: '#/components/parameters/SkillId'
    get:
      tags:
        - Admin Skills
      operationId: getSkill
      responses:
        '200':
          $ref: '#/components/responses/AdminSkill'
        '404':
          $ref: '#/components/responses/Problem-2'
    patch:
      tags:
        - Admin Skills
      operationId: updateSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillPatch'
      responses:
        '200':
          $ref: '#/components/responses/AdminSkill'
    delete:
      tags:
        - Admin Skills
      operationId: deleteSkill
      responses:
        '204':
          description: Skill withdrawn
  /aep/v1/admin/skills/{skillId}/versions:
    post:
      tags:
        - Admin Skills
      operationId: uploadSkillVersion
      parameters:
        - $ref: '#/components/parameters/SkillId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - version
                - package
              properties:
                version:
                  type: string
                releaseNotes:
                  type: string
                package:
                  type: string
                  format: binary
      responses:
        '201':
          description: Draft version uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillVersion'
  /aep/v1/admin/skills/{skillId}/versions/{version}/publish:
    post:
      tags:
        - Admin Skills
      operationId: publishSkillVersion
      parameters:
        - $ref: '#/components/parameters/SkillId'
        - $ref: '#/components/parameters/Version'
      responses:
        '200':
          description: Version published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillVersion'
  /aep/v1/admin/skill-assignments:
    get:
      tags:
        - Admin Skills
      operationId: listSkillAssignments
      responses:
        '200':
          $ref: '#/components/responses/AssignmentList'
    post:
      tags:
        - Admin Skills
      operationId: createSkillAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillAssignmentWrite'
      responses:
        '201':
          $ref: '#/components/responses/Assignment'
  /aep/v1/admin/skill-assignments/{assignmentId}:
    delete:
      tags:
        - Admin Skills
      operationId: deleteSkillAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '204':
          description: Assignment removed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Problem-2:
      type: object
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
        requestId:
          type: string
...
    SkillSyncItem:
      type: object
      required:
        - skillId
        - version
        - action
        - status
      properties:
        skillId:
          type: string
        version:
          type: string
        action:
          type: string
          enum:
            - install
            - update
            - remove
            - unchanged
        status:
          type: string
          enum:
            - success
            - failed
        errorCode:
          type: string
        message:
          type: string
    SkillSyncResult:
      type: object
      required:
        - manifestRevision
        - startedAt
        - completedAt
        - status
        - items
      properties:
        manifestRevision:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - success
            - partial
            - failed
        items:
          type: array
          items:
            $ref: '#/components/schemas/SkillSyncItem'
...
    SkillVersion:
      type: object
      required:
        - version
        - state
        - sha256
        - size
        - createdAt
      properties:
        version:
          type: string
        state:
          type: string
          enum:
            - draft
            - published
            - withdrawn
        sha256:
          type: string
          pattern: ^[a-f0-9]{64}$
        size:
          type: integer
          minimum: 0
        releaseNotes:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminSkill:
      type: object
      required:
        - id
        - name
        - description
        - state
        - versions
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        state:
          type: string
          enum:
            - active
            - withdrawn
        versions:
          type: array
          items:
            $ref: '#/components/schemas/SkillVersion'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminSkillList:
      type: object
      required:
        - skills
      properties:
        skills:
          type: array
          items:
            $ref: '#/components/schemas/AdminSkill'
    SkillWrite:
      type: object
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
    SkillPatch:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        description:
          type: string
        state:
          type: string
          enum:
            - active
            - withdrawn
    Subject:
      type: object
      required:
        - type
        - id
      properties:
        type:
          type: string
          enum:
            - user
            - role
            - team
        id:
          type: string
    Assignment:
      type: object
      required:
        - id
        - resourceType
        - resourceId
        - subject
        - createdAt
      properties:
        id:
          type: string
        resourceType:
          type: string
          enum:
            - skill
            - credential
            - model
        resourceId:
          type: string
        subject:
          $ref: '#/components/schemas/Subject'
        createdAt:
          type: string
          format: date-time
    AssignmentList:
      type: object
      required:
        - assignments
      properties:
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/Assignment'
    SkillAssignmentWrite:
      type: object
      required:
        - skillId
        - subject
      properties:
        skillId:
          type: string
        subject:
          $ref: '#/components/schemas/Subject'
...
  responses:
...
    Problem-2:
      description: AEP Problem Details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem-2'
    AdminSkill:
      description: Skill resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminSkill'
    AssignmentList:
      description: Assignment list
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AssignmentList'
    Assignment:
      description: Assignment resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Assignment'
...
  parameters:
...
    SkillId:
      name: skillId
      in: path
      required: true
      schema:
        type: string
    Version:
      name: version
      in: path
      required: true
      schema:
        type: string
...
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
...